Restructure db

+++ Dominio/Entidades/Usuario.cs
@@ -15,7 +15,7 @@ namespace Dominio.Entidades
 
         public Rol Rol { get; set; }
 
-        public long PersonaId { get; set; }
+        public int PersonaId { get; set; }
 
         public DateTime Fecha { get; set; }
